# frontend/lib/api/openapi.yml
openapi: 3.0.3
info:
  title: Kuyash Place Restaurant API
  version: 1.0.0
  description: |-
    Backend API for Kuyash Place Restaurant.

    **All money values are integer kobo** wrapped in a `{amount, currency, display}` object. Clients render `display` and never perform arithmetic on `amount`.
paths:
...
  /api/v1/accounts/me/:
    get:
      operationId: accounts_me_retrieve
      description: Read, update or erase the current user.
      summary: Current profile
      tags:
      - accounts
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
          description: ''
    patch:
      operationId: accounts_me_partial_update
      description: Read, update or erase the current user.
      summary: Update profile
      tags:
      - accounts
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedProfileRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedProfileRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedProfileRequest'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
          description: ''
    delete:
      operationId: accounts_me_destroy
      description: Read, update or erase the current user.
      summary: Erase account (NDPR)
      tags:
      - accounts
      security:
      - cookieAuth: []
      responses:
        '204':
          description: No response body
...
components:
  schemas:
...
    PatchedProfileRequest:
      type: object
      description: |-
        Read/write profile.

        ``email`` is deliberately read-only. Changing it must re-verify the new
        address and notify the old one (AS-5); until that flow exists in Phase 2,
        the safe behaviour is to refuse the change rather than allow an unverified one.
      properties:
        full_name:
          type: string
          description: The UI collects a single name field, not first/last. Optional
            at the model level because social sign-in may not supply one; the signup
            serializer requires it.
          maxLength: 150
        phone:
          type: string
          description: International format. Required before placing a delivery order.
          pattern: ^\+?[1-9]\d{7,14}$
          maxLength: 20
        date_of_birth:
          type: string
          format: date
          nullable: true
        marketing_opt_in:
          type: boolean
...
    Profile:
      type: object
      description: |-
        Read/write profile.

        ``email`` is deliberately read-only. Changing it must re-verify the new
        address and notify the old one (AS-5); until that flow exists in Phase 2,
        the safe behaviour is to refuse the change rather than allow an unverified one.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        email:
          type: string
          format: email
          readOnly: true
        full_name:
          type: string
          description: The UI collects a single name field, not first/last. Optional
            at the model level because social sign-in may not supply one; the signup
            serializer requires it.
          maxLength: 150
        phone:
          type: string
          description: International format. Required before placing a delivery order.
          pattern: ^\+?[1-9]\d{7,14}$
          maxLength: 20
        is_email_verified:
          type: boolean
          readOnly: true
        date_of_birth:
          type: string
          format: date
          nullable: true
        marketing_opt_in:
          type: boolean
        date_joined:
          type: string
          format: date-time
          readOnly: true
      required:
      - date_joined
      - email
      - id
      - is_email_verified
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: kuyash_session